backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
